+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 - feat: added extended tx info in search response [DV-3827]
+- feat: added endpoint to retrieve extended information about store enabled currencies [DV-3857]
 
 ## [0.9.13] - 2025-11-19
 - Fix `FindLastWalletTransactions` sql query  [DV-3808]